requirements.txt
asgiref==3.4.1
astroid==2.7.3
colorama==0.4.4
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
platformdirs==2.3.0
psycopg2==2.9.1
pylint==2.10.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
wrapt==1.12.1
